+++ movieland-backend/src/main/java/com/phil/movieland/rest/controller/SeatController.java
@@ -21,7 +21,7 @@ public class SeatController {
     }
 
     @GetMapping("/reservations/{resId}")
-    public List<Seat> getSeatsOfReservation(@PathVariable Long resId) {
+    public List<Seat> getSeatsOfReservation(@PathVariable Integer resId) {
         return reservationService.getAllSeatsOfReservation(resId);
     }
